backend/requirements.txt
pymongo==4.6.0
pandas==2.1.4
openpyxl==3.1.2
bcrypt==4.1.2
python-dotenv==1.1.1
